+++ .readthedocs.yml
@@ -12,3 +12,4 @@ sphinx:
 python:
   install:
     - requirements: docs/requirements.txt
+    - requirements: docs/html/requirements.txt
