+++ docs/html/conf.py
@@ -1,15 +1,15 @@
 """Sphinx configuration file for pip's documentation."""
 
-import glob
 import os
 import pathlib
-import re
 import sys
-from typing import List, Tuple
 
-# Add the docs/ directory to sys.path, because pip_sphinxext.py is there.
+# Add the docs/ directory to sys.path to load the common config,
+# and pip_sphinxext.py
 docs_dir = os.path.dirname(os.path.dirname(__file__))
 sys.path.insert(0, docs_dir)
+
+from common_conf import copyright, project, release, version  # noqa: E402, F401
 
 # -- General configuration ------------------------------------------------------------
 
@@ -27,27 +27,6 @@ extensions = [
     "sphinx_inline_tabs",
     "sphinxcontrib.towncrier",
 ]
-
-# General information about the project.
-project = "pip"
-copyright = "The pip developers"
-
-# Find the version and release information.
-# We have a single source of truth for our version number: pip's __init__.py file.
-# This next bit of code reads from it.
-file_with_version = os.path.join(docs_dir, "..", "src", "pip", "__init__.py")
-with open(file_with_version) as f:
-    for line in f:
-        m = re.match(r'__version__ = "(.*)"', line)
-        if m:
-            __version__ = m.group(1)
-            # The short X.Y version.
-            version = ".".join(__version__.split(".")[:2])
-            # The full version, including alpha/beta/rc tags.
-            release = __version__
-            break
-    else:  # AKA no-break
-        version = release = "dev"
 
 print("pip version:", version)
 print("pip release:", release)
@@ -94,43 +73,6 @@ html_title = f"{project} documentation v{release}"
 # Disable the generation of the various indexes
 html_use_modindex = False
 html_use_index = False
-
-# -- Options for Manual Pages ---------------------------------------------------------
-
-
-# List of manual pages generated
-def determine_man_pages() -> List[Tuple[str, str, str, str, int]]:
-    """Determine which man pages need to be generated."""
-
-    def to_document_name(path: str, base_dir: str) -> str:
-        """Convert a provided path to a Sphinx "document name"."""
-        relative_path = os.path.relpath(path, base_dir)
-        root, _ = os.path.splitext(relative_path)
-        return root.replace(os.sep, "/")
-
-    # Crawl the entire man/commands/ directory and list every file with appropriate
-    # name and details.
-    man_dir = os.path.join(docs_dir, "man")
-    raw_subcommands = glob.glob(os.path.join(man_dir, "commands/*.rst"))
-    if not raw_subcommands:
-        raise FileNotFoundError(
-            "The individual subcommand manpages could not be found!"
-        )
-
-    retval = [
-        ("index", "pip", "package manager for Python packages", "pip developers", 1),
-    ]
-    for fname in raw_subcommands:
-        fname_base = to_document_name(fname, man_dir)
-        outname = "pip-" + fname_base.split("/")[1]
-        description = "description of {} command".format(outname.replace("-", " "))
-
-        retval.append((fname_base, outname, description, "pip developers", 1))
-
-    return retval
-
-
-man_pages = determine_man_pages()
 
 # -- Options for sphinx_copybutton ----------------------------------------------------
