+++ docs/html/requirements.txt
@@ -1,0 +1,8 @@
+# currently incompatible with sphinxcontrib-towncrier
+# https://github.com/sphinx-contrib/sphinxcontrib-towncrier/issues/92
+towncrier < 24
+furo
+myst_parser
+sphinx-copybutton
+sphinx-inline-tabs
+sphinxcontrib-towncrier >= 0.2.0a0
