+++ docs/requirements.txt
@@ -1,12 +1,4 @@
 sphinx ~= 7.0
-# currently incompatible with sphinxcontrib-towncrier
-# https://github.com/sphinx-contrib/sphinxcontrib-towncrier/issues/92
-towncrier < 24
-furo
-myst_parser
-sphinx-copybutton
-sphinx-inline-tabs
-sphinxcontrib-towncrier >= 0.2.0a0
 
 # `docs.pipext` uses pip's internals to generate documentation. So, we install
 # the current directory to make it work.
